Script automatization
Now it is not necessary to edit the script to set the configurations of the train and test. In this version it is possible to run it via command line and pass the arguments suh as error_cost, number of trees, columns to use, local path and training partition percentual.