New version of decision three: UNBALANCED
In this version, actually a beta version, the data for training is as unbalanced as the original dataset. 
It uses a subset due to fit the machine where it will be runned (8GB RAM).

The "behavior" of this subset need to be checked. It must be representative considering the original dataset.